Add an ability to Y-flip images on import.
While the encoder-provided option to flip the image during encoding is
everything needed when one controls the whole pipeline, it's of no use
for general image viewers that have no power over how the input was
generated. Having the chance to configure Y flipping during the
transcode makes it easier to integrate, as the application doesn't need
to worry about patching texture coordinates or doing texture coordinate
transform in a shader.

While flipping the compressed image post-transcode is possible, it's a
non-trivial amount of work, considering the broad range of target
formats supported. On the other hand, when done directly in Basis,
flipping can be done on the internal ETC1S representation:

 1. patching the output pointer and making row stride negative, which
    has very little impact on existing code;
 2. and bit-flipping the selector bits using a modified 3-operation bit
    reverse trick.

Assuming the encoder's output is reasonably invariant to image flip,
this new code path could be tested by comparing these two outputs for
equality:

- flipped PNG -> basis -> target format -> RGBA
- PNG -> basis -> flipped target format -> RGBA

This was tested on ETC1/2, BC1-7, ASTC and RGBA8 output, however it
doesn't handle flipping for PVRTC textures, as there's cross-block
modulation and I don't have deep enough understanding of the format to
be able to code that. Similarly for FXT1 due to a different block size.